Time: 17 ms (83.26%) | Memory: 56.3 MB (78.94%) - LeetSync

+++ 1605-minimum-number-of-days-to-make-m-bouquets/minimum-number-of-days-to-make-m-bouquets.java
@@ -1,7 +1,7 @@
 class Solution {
     public int minDays(int[] bloomDay, int m, int k) {
-        long n = bloomDay.length;
-        long z = m*k;
+        int n = bloomDay.length;
+        int z = m*k;
         if(n < z){
             return -1;
         }
